requirements.txt
# ============================================================
# requirements.txt — Autonomous Driving Multimodal SSL Pipeline
# ============================================================
# Install with: pip install -r requirements.txt
# For Colab: !pip install -r requirements.txt
# ============================================================

# Core deep learning
torch>=2.1.0
torchvision>=0.16.0

# Data processing
numpy>=1.24.0
Pillow>=10.0.0
open3d>=0.18.0
opencv-python>=4.8.0

# Training utilities
tqdm>=4.65.0
scikit-learn>=1.3.0
pandas>=2.0.0
scipy>=1.11.0

# Visualisation
matplotlib>=3.7.0

# REST API server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.0.0
